requirements.txt
streamlit
pandas
matplotlib
seaborn
langchain
langchain-community
langchain-google-genai
ollama
langchain-experimental
tabulate
python-docx